Fix: print_job-Migration 0142 (gsl_id INT statt BIGINT) - MariaDB errno 150
CREATE TABLE print_job faultete im Prod-Deploy mit
(1005, ... errno 150 "Foreign key constraint is incorrectly formed").
Ursache: print_job.gsl_id war BIGINT, die referenzierte Spalte
major_incident.id ist aber INT (sa.Integer, Migration 0037). MySQL/MariaDB
verlangt exakte Typgleichheit fuer FK-Spalten.

gsl_id auf INT (Migration) bzw. Integer (Modell PrintJob) korrigiert. Alle
uebrigen print_job-FKs referenzieren BIGINT-PKs und sind unveraendert.

+++ alembic/versions/0142_print_job.py
@@ -24,7 +24,9 @@ def upgrade() -> None:
             `source`          VARCHAR(20) NOT NULL DEFAULT 'manual',
             `rule_id`         BIGINT NULL,
             `incident_id`     BIGINT NULL,
-            `gsl_id`          BIGINT NULL,
+            -- major_incident.id ist INT (nicht BIGINT) — gsl_id muss exakt den Typ der
+            -- referenzierten Spalte haben, sonst MySQL/MariaDB errno 150 (FK incorrectly formed).
+            `gsl_id`          INT NULL,
             `objekt_id`       BIGINT NULL,
             `document_type`   VARCHAR(40) NOT NULL,
             `artifact_ref`    VARCHAR(120) NULL,
